fix build error on entity.cpp

+++ src/Entity.cpp
@@ -1,6 +1,6 @@
 #include "Entity.hpp"
 
-void Entity::set_pos(int x, int y) {
+void Entity::set_pos(double x, double y) {
   posX = x;
   posY = y;
 }
